% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{brain}
\alias{brain}
\title{Brain RNA-Seq data for both human and mouse.}
\format{Large matrix with 15041 rows and 62 columns}
\source{
Zhang, Ye, et al. "Purification and characterization of progenitor and mature human astrocytes reveals transcriptional and functional differences with mouse." Neuron 89.1 (2016): 37-53.
}
\usage{
data(brain)
}
\description{
Brain RNA-Seq data for both human and mouse.
}
\keyword{datasets}
