% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{derif}
\alias{derif}
\title{Column gradient for the loss function based on Pearson correlation distance (faster Rcpp version)}
\usage{
derif(c, w, d, core, idx)
}
\arguments{
\item{c}{The original p*n batch effect data with n subjects and p RNA-seq measurements.}

\item{w}{An initial n*n weight matrix to conduct linear transformation. Default to be identity matrix if not specified.}

\item{d}{The n*n distance matrix obtained by QuantNorm.}

\item{core}{Pre-calculated component from main algorithm in order to save computation time.}

\item{idx}{Random selected columns input from the main algorithm.}
}
\description{
A specific implementation of column gradient to be used in the main coordinate gradient descent algorithm scBatchCpp
}
\author{
Teng Fei. Email: tfei@emory.edu
}
