% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODE.r
\docType{data}
\name{ENCODE}
\alias{ENCODE}
\title{Normalized ENCODE raw counts data for both human and mouse.}
\format{Large matrix with 10309 rows and 26 columns}
\source{
Reproduced according to Gilad, Yoav, and Orna Mizrahi-Man. "A reanalysis of mouse ENCODE comparative gene expression data." F1000Research 4 (2015).
}
\usage{
data(ENCODE)
}
\description{
Normalized ENCODE raw counts data for both human and mouse.
}
\keyword{datasets}
