% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scBatchCpp}
\alias{scBatchCpp}
\title{Correct scRNA-seq count matrix subject to batch effects by sample distance matrix correction}
\usage{
scBatchCpp(c, w, d, m, max, step, tol, derif)
}
\arguments{
\item{c}{The original p*n batch effect data with n subjects and p RNA-seq measurements.}

\item{w}{An initial n*n weight matrix to conduct linear transformation. Default to be identity matrix if not specified.}

\item{d}{The n*n distance matrix obtained by QuantNorm.}

\item{m}{Number of groups to be divided for coordinate gradient descent. 1 < m <= n. Default to be 0.1n if not specified.}

\item{max}{Maximum number of the iteration if the tolerance is not reached.}

\item{step}{Step size of the gradient descent algorithm.}

\item{tol}{Stopping criteria of the algorithm. The algorithm stops if the step size is smaller than tol.}

\item{derif}{Function to compute gradient of the loss function.}
}
\value{
Returns the corrected count matrix.
}
\description{
As an extension of package 'QuantNorm', which corrects the distance matrix to obtain reasonable sample pattern, this package use gradient descent algorithm to correct count matrix by linear transformation.
}
\references{
Fei et al (2018), Mitigating the adverse impact of batch effects in sample pattern detection, Bioinformatics, <https://doi.org/10.1093/bioinformatics/bty117>.
}
\author{
Teng Fei. Email: tfei@emory.edu
}
