% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scBatchCpp}
\alias{scBatchCpp}
\title{Correct RNA-seq count matrix subject to batch effects by sample distance matrix correction (faster Rcpp version)}
\usage{
scBatchCpp(c, w, d, m, max, step, tol, derif, verbose)
}
\arguments{
\item{c}{The original p*n batch effect data with n subjects and p RNA-seq measurements.}

\item{w}{An initial n*n weight matrix to conduct linear transformation. An usual choice is the identity matrix if not specified.}

\item{d}{The n*n distance matrix obtained by QuantNorm.}

\item{m}{Number of groups to be divided for coordinate gradient descent. 1 < m <= n.}

\item{max}{Maximum number of the iteration if the tolerance is not reached.}

\item{step}{Step size of the gradient descent algorithm.}

\item{tol}{Stopping criteria of the algorithm. The algorithm stops if the step size is smaller than tol.}

\item{derif}{Function to compute gradient of the loss function. For Pearson distance, use scBatch::derif.}

\item{verbose}{Whether output the loss function after each iteration.}
}
\value{
Returns the corrected count matrix.
}
\description{
Use gradient descent algorithm to correct count matrix by linear transformation based on known sample matrix correction.
}
\references{
Fei & Yu (2020). scBatch: batch-effect correction of RNA-seq data through sample distance matrix adjustment. Bioinformatics, 36(10), 3115-3123. <https://doi.org/10.1093/bioinformatics/btaa097>.
}
\author{
Teng Fei. Email: tfei@emory.edu
}
